refactor: rename VideoAbilities to MediaAbilities, abilities to upload-media/validate-media
Promotes the upload and validate abilities to media-level since the
MediaValidator abstraction already handles both image and video types:

- datamachine/upload-video  → datamachine/upload-media
- datamachine/validate-video → datamachine/validate-media
- datamachine/video-metadata stays as-is (ffprobe is video-specific)

upload-media and validate-media auto-detect media type from file extension
(VideoValidator::is_video_extension) and use the correct MediaValidator
subclass. Callers can force media_type='image' or 'video' on validate-media.

Response now includes media_type field ('image' or 'video') so consumers
know which type was detected.
